+++ K-Means/Main.cpp
@@ -4,7 +4,7 @@ int main(int argc, char* argv[])
 {
 	clock_t start, end;
 	int rank, numberOfProcesses, N, K, LIMIT, myNumberOfPoints, iterationNumber = 0, pathSize;
-	float dT, time, T, QM, q;
+	double dT, time, T, QM, q;
 	bool anotherIteration, isMovedPoint = NULL;
 	point_t *allPoints, *myPoints, *clusters, *incPoints = NULL, *previousIncPoints = NULL;
 	char path[PATH_SIZE];
